Customizing the validation error

+++ src/main/java/api/restaurant/entity/Category.java
@@ -22,7 +22,7 @@ public class Category implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 
 	@ManyToMany(mappedBy = "categories")
@@ -32,7 +32,7 @@ public class Category implements Serializable {
 
 	}
 
-	public Category(Integer id, String name) {
+	public Category(Long id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
